refactor: import parse_date directly
